@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:sonnen_rennt/constants/color.dart';
 import 'package:sonnen_rennt/constants/theme.dart';
 import 'package:sonnen_rennt/root.dart';
 
@@ -10,9 +12,15 @@ class SunRunApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
+
+    SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle.dark.copyWith(
+      systemNavigationBarColor: SunRunColors.djk_bg_darker
+    ));
+
     return MaterialApp(
       title: 'SunRun',
       theme: sunRunThemeBasic(),
+      debugShowCheckedModeBanner: false,
       // ThemeData(
       //   primarySwatch: Colors.orange,
       // )
@@ -22,9 +30,9 @@ class SunRunApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, this.title}) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   MyHomePageState createState() => MyHomePageState();
