@@ -27,7 +27,7 @@ class _GroupListScreenState extends State<GroupListScreen> {
 
   _GroupListScreenState({this.onlyLoggedInUsers});
 
-  bool onlyLoggedInUsers = true;
+  bool? onlyLoggedInUsers = true;
 
   void updateList() {
     setState(() {});
@@ -36,7 +36,7 @@ class _GroupListScreenState extends State<GroupListScreen> {
   @override
   Widget build(BuildContext context) {
     return SunRunBaseWidget(
-      title: onlyLoggedInUsers ? "Meine Gruppen" : "Alle Gruppen",
+      title: onlyLoggedInUsers! ? "Meine Gruppen" : "Alle Gruppen",
       child: Column(
         children: [
           Padding(
@@ -53,16 +53,18 @@ class _GroupListScreenState extends State<GroupListScreen> {
           ),
           SizedBox(height: 14,),
           FutureBuilder(
-              future: onlyLoggedInUsers
+              future: onlyLoggedInUsers!
                   ? GroupApi.groupListUser()
                   : GroupApi.groupList(),
-              builder: (context, snap) {
+              builder: (
+                  BuildContext context,
+                  AsyncSnapshot<GroupListResult> snap) {
                 // TODO waiting & error
 
-                if(snap.hasError ||
-                    (snap.hasData && snap.data.type != GroupListResultType.SUCCESS_200)) {
+                if(snap.hasError || (snap.hasData && snap.data != null
+                    && snap.data!.type != GroupListResultType.SUCCESS_200)) {
                   return SrErrorWidget(
-                    description: snap.data.detail,
+                    description: snap.data!.detail,
                   );
                 }
 
@@ -70,8 +72,9 @@ class _GroupListScreenState extends State<GroupListScreen> {
                   return SrWaitingWidget();
                 }
 
-                if(snap.hasData && snap.data.type == GroupListResultType.SUCCESS_200) {
-                  GroupListResult res = snap.data;
+                if(snap.hasData && snap.data != null &&
+                    snap.data!.type == GroupListResultType.SUCCESS_200) {
+                  GroupListResult res = snap.data!;
                   List<Group> groups = res.groups;
 
                   return Container(
@@ -93,30 +96,28 @@ class _GroupListScreenState extends State<GroupListScreen> {
                   );
                 }
                 return Text("ERROR");
-
-
-                if(snap.hasData && !snap.hasError) {
-                  return Container(
-                    height: MediaQuery.of(context).size.height * 0.85,
-                    child: ListView(
-                      shrinkWrap: true,
-                      children: new List.generate(snap.data.length+1, (index) {
-                        if (index == snap.data.length) {
-                          return SizedBox(height: 200,);
-                        }
-                        Group group = snap.data[index];
-                        print("route: " + group.toString());
-                        return GroupListCardWidget(
-                          group: group,
-                          onlyLoggedInUsers: onlyLoggedInUsers,
-                        );
-                      }).toList(),
-                    ),
-                  );
-                }
-                return Container(
-                  height: 1,
-                );
+                // if(snap.hasData && !snap.hasError) {
+                //   return Container(
+                //     height: MediaQuery.of(context).size.height * 0.85,
+                //     child: ListView(
+                //       shrinkWrap: true,
+                //       children: new List.generate(snap.data.length+1, (index) {
+                //         if (index == snap.data.length) {
+                //           return SizedBox(height: 200,);
+                //         }
+                //         Group? group = snap.data[index];
+                //         print("route: " + group.toString());
+                //         return GroupListCardWidget(
+                //           group: group,
+                //           onlyLoggedInUsers: onlyLoggedInUsers,
+                //         );
+                //       }).toList(),
+                //     ),
+                //   );
+                // }
+                // return Container(
+                //   height: 1,
+                // );
               }
           )
         ],
@@ -129,8 +130,8 @@ class GroupListCardWidget extends StatefulWidget {
 
   GroupListCardWidget({this.group, this.onlyLoggedInUsers});
 
-  Group group;
-  bool onlyLoggedInUsers = true;
+  Group? group;
+  bool? onlyLoggedInUsers = true;
 
   @override
   _GroupListCardWidgetState createState() =>
@@ -144,8 +145,8 @@ class _GroupListCardWidgetState extends State<GroupListCardWidget> {
 
   _GroupListCardWidgetState({this.group, this.onlyLoggedInUsers});
 
-  Group group;
-  bool onlyLoggedInUsers = true;
+  Group? group;
+  bool? onlyLoggedInUsers = true;
 
   bool toggle = false;
 
@@ -173,7 +174,7 @@ class _GroupListCardWidgetState extends State<GroupListCardWidget> {
                   mainAxisAlignment: MainAxisAlignment.start,
                   children: [
                     Flexible(
-                        child: Text(group.name,
+                        child: Text(group!.name!,
                           style: TextStyle(color: SunRunColors.djk_heading, fontSize: 18, fontWeight: FontWeight.bold),)
                     ),
                   ],
@@ -182,13 +183,13 @@ class _GroupListCardWidgetState extends State<GroupListCardWidget> {
                 Row(
                   children: [
                     Text("Score: ", style: TextStyle(color: SunRunColors.djk_heading, fontWeight: FontWeight.bold),),
-                    Text(group.score.toStringAsFixed(2) + " Punkte", style: TextStyle(color: SunRunColors.djk_heading),),
+                    Text(group!.score!.toStringAsFixed(2) + " Punkte", style: TextStyle(color: SunRunColors.djk_heading),),
                   ],
                 ),
                 SizedBox(height: 12,),
                 Builder(
                   builder: (ctx) {
-                    if (group.description != null && group.description.length > 0) {
+                    if (group!.description != null && group!.description!.length > 0) {
                       return Column(
                         crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
@@ -197,7 +198,7 @@ class _GroupListCardWidgetState extends State<GroupListCardWidget> {
                           SizedBox(height: 8,),
                           Container(
                             width: MediaQuery.of(context).size.width * 0.75,
-                            child: Text(group.description, style: TextStyle(color: SunRunColors.djk_heading), maxLines: 10,),
+                            child: Text(group!.description!, style: TextStyle(color: SunRunColors.djk_heading), maxLines: 10,),
                           )
                         ],
                       );
@@ -217,21 +218,21 @@ class _GroupListCardWidgetState extends State<GroupListCardWidget> {
                           Row(
                             children: [
                               Text("Anzahl Aktivitäten: ", style: TextStyle(color: SunRunColors.djk_heading, fontWeight: FontWeight.bold),),
-                              Text(group.run_count.toString(), style: TextStyle(color: SunRunColors.djk_heading),),
+                              Text(group!.run_count.toString(), style: TextStyle(color: SunRunColors.djk_heading),),
                             ],
                           ),
                           SizedBox(height: 12,),
                           Row(
                             children: [
                               Text("Dauer: ", style: TextStyle(color: SunRunColors.djk_heading, fontWeight: FontWeight.bold),),
-                              Text(group.sumDurationFormatted, style: TextStyle(color: SunRunColors.djk_heading),),
+                              Text(group!.sumDurationFormatted, style: TextStyle(color: SunRunColors.djk_heading),),
                             ],
                           ),
                           SizedBox(height: 12,),
                           Row(
                             children: [
                               Text("Teilnehmerzahl: ", style: TextStyle(color: SunRunColors.djk_heading, fontWeight: FontWeight.bold),),
-                              Text(group.num_participants.toString(), style: TextStyle(color: SunRunColors.djk_heading),),
+                              Text(group!.num_participants.toString(), style: TextStyle(color: SunRunColors.djk_heading),),
                             ],
                           ),
                           SizedBox(height: 12,),
@@ -242,14 +243,14 @@ class _GroupListCardWidgetState extends State<GroupListCardWidget> {
                                 children: [
                                   Icon(Icons.directions_bike, color: Colors.white,),
                                   SizedBox(width: 10,),
-                                  Text(group.sum_distance_bike.toString() + "km", style: TextStyle(color: SunRunColors.djk_heading),),
+                                  Text(group!.sum_distance_bike.toString() + "km", style: TextStyle(color: SunRunColors.djk_heading),),
                                 ],
                               ),
                               Row(
                                 children: [
                                   Icon(Icons.electric_bike, color: Colors.white,),
                                   SizedBox(width: 10,),
-                                  Text(group.sum_distance_ebike.toString() + "km", style: TextStyle(color: SunRunColors.djk_heading),),
+                                  Text(group!.sum_distance_ebike.toString() + "km", style: TextStyle(color: SunRunColors.djk_heading),),
                                 ],
                               ),
                             ],
@@ -262,14 +263,14 @@ class _GroupListCardWidgetState extends State<GroupListCardWidget> {
                                 children: [
                                   Icon(Icons.directions_run, color: Colors.white,),
                                   SizedBox(width: 10,),
-                                  Text(group.sum_distance_run.toString() + "km", style: TextStyle(color: SunRunColors.djk_heading),),
+                                  Text(group!.sum_distance_run.toString() + "km", style: TextStyle(color: SunRunColors.djk_heading),),
                                 ],
                               ),
                               Row(
                                 children: [
                                   Icon(Icons.directions_walk, color: Colors.white,),
                                   SizedBox(width: 10,),
-                                  Text(group.sum_distance_walk.toString() + "km", style: TextStyle(color: SunRunColors.djk_heading),),
+                                  Text(group!.sum_distance_walk.toString() + "km", style: TextStyle(color: SunRunColors.djk_heading),),
                                 ],
                               ),
                             ],
