@@ -11,26 +11,26 @@ class DJKRoute {
     this.description, this.link, this.creatorId
   });
 
-  int id;
-  String title;
-  double distance;
-  double elevationGain;
-  String description;
-  String link;
-  int creatorId;
+  int? id;
+  String? title;
+  double? distance;
+  double? elevationGain;
+  String? description;
+  String? link;
+  int? creatorId;
 
-  Map<String, String> toMap() {
-    Map<String, String> map = <String, String> {
+  Map<String, String?> toMap() {
+    Map<String, String?> map = <String, String?> {
       'title': title,
       'distance': distance.toString(),
       'elevation_gain': elevationGain.toString(),
     };
 
-    if (description != null && description.isNotEmpty) {
+    if (description != null && description!.isNotEmpty) {
       map.addAll({ 'description': description });
     }
 
-    if (link != null && link.length > 0) {
+    if (link != null && link!.length > 0) {
       map.addAll({ 'link': link.toString() });
     }
 
@@ -46,21 +46,21 @@ class DJKRoute {
       double distance = double.parse(json['distance'].toString());
       double elevationGain = double.parse(json['elevation_gain'].toString());
 
-      String title = json['title'];
+      String? title = json['title'];
       if (json['title'] == null || json['title'].toString() == "null") {
         title = null;
       } else {
         title = json['title'].toString();
       }
 
-      String description = json['description'];
+      String? description = json['description'];
       if (json['description'] == null || json['description'].toString() == "null") {
         description = null;
       } else {
         description = json['description'].toString();
       }
 
-      String link = json['link'];
+      String? link = json['link'];
       if (json['link'] == null || json['link'].toString() == "null") {
         link = null;
       } else {
@@ -73,10 +73,10 @@ class DJKRoute {
       );
     } on Exception catch(ex) {
       print(ex.toString());
-      return null;
+      throw ex;
     } on Error catch(err) {
       print(err.toString());
-      return null;
+      throw err;
     }
   }
 
