@@ -7,9 +7,9 @@ class CardCategory extends StatelessWidget {
       this.img = "https://via.placeholder.com/250",
       this.tap = defaultFunc});
 
-  final String img;
+  final String? img;
   final Function tap;
-  final String title;
+  final String? title;
 
   static void defaultFunc() {
     print("the function works!");
@@ -21,7 +21,7 @@ class CardCategory extends StatelessWidget {
         height: 252,
         width: null,
         child: GestureDetector(
-          onTap: tap,
+          onTap: tap as void Function()?,
           child: Card(
               elevation: 0.4,
               shape: RoundedRectangleBorder(
@@ -31,7 +31,7 @@ class CardCategory extends StatelessWidget {
                     decoration: BoxDecoration(
                         borderRadius: BorderRadius.all(Radius.circular(6.0)),
                         image: DecorationImage(
-                          image: NetworkImage(img),
+                          image: NetworkImage(img!),
                           fit: BoxFit.cover,
                         ))),
                 Container(
@@ -39,7 +39,7 @@ class CardCategory extends StatelessWidget {
                         color: Colors.black45,
                         borderRadius: BorderRadius.all(Radius.circular(6.0)))),
                 Center(
-                  child: Text(title,
+                  child: Text(title!,
                       style: TextStyle(
                           color: SunRunColors.white,
                           fontWeight: FontWeight.w600,
