@@ -10,9 +10,9 @@ class CardHorizontal extends StatelessWidget {
       this.tap = defaultFunc});
 
   final String cta;
-  final String img;
+  final String? img;
   final Function tap;
-  final String title;
+  final String? title;
 
   static void defaultFunc() {
     print("the function works!");
@@ -23,7 +23,7 @@ class CardHorizontal extends StatelessWidget {
     return Container(
         height: 130,
         child: GestureDetector(
-          onTap: tap,
+          onTap: tap as void Function()?,
           child: Card(
             elevation: 3,
             shadowColor: SunRunColors.muted.withOpacity(0.22),
@@ -39,7 +39,7 @@ class CardHorizontal extends StatelessWidget {
                               topLeft: Radius.circular(4.0),
                               bottomLeft: Radius.circular(4.0)),
                           image: DecorationImage(
-                            image: NetworkImage(img),
+                            image: NetworkImage(img!),
                             fit: BoxFit.cover,
                           ))),
                 ),
@@ -51,7 +51,7 @@ class CardHorizontal extends StatelessWidget {
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
-                          Text(title,
+                          Text(title!,
                               style: TextStyle(
                                   color: SunRunColors.text, fontSize: 12)),
                           Text(cta,
