@@ -3,9 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:sonnen_rennt/constants/color.dart';
 
 class DrawerTile extends StatelessWidget {
-  final String title;
-  final IconData icon;
-  final Function onTap;
+  final String? title;
+  final IconData? icon;
+  final Function? onTap;
   final bool isSelected;
   final Color iconColor;
 
@@ -19,7 +19,7 @@ class DrawerTile extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: onTap,
+      onTap: onTap as void Function()?,
       child: Container(
           margin: EdgeInsets.only(top: 6),
           padding: EdgeInsets.all(16),
@@ -44,7 +44,7 @@ class DrawerTile extends StatelessWidget {
                       : SunRunColors.white.withOpacity(1.0)),
               Padding(
                 padding: const EdgeInsets.only(left: 8.0),
-                child: Text(title,
+                child: Text(title!,
                     style: TextStyle(
                         letterSpacing: .3,
                         fontSize: 15,
