@@ -6,11 +6,11 @@ import 'package:sonnen_rennt/constants/Theme.dart';
 import 'package:sonnen_rennt/constants/color.dart';
 
 class ProductCarousel extends StatefulWidget {
-  final List<Map<String, String>> imgArray;
+  final List<Map<String, String>>? imgArray;
 
   const ProductCarousel({
-    Key key,
-    @required this.imgArray,
+    Key? key,
+    required this.imgArray,
   }) : super(key: key);
 
   @override
@@ -23,7 +23,7 @@ class _ProductCarouselState extends State<ProductCarousel> {
   @override
   Widget build(BuildContext context) {
     return CarouselSlider(
-      items: widget.imgArray
+      items: widget.imgArray!
           .map((item) => Container(
                 child: Column(
                   children: [
@@ -42,7 +42,7 @@ class _ProductCarouselState extends State<ProductCarousel> {
                           child: ClipRRect(
                             borderRadius: BorderRadius.circular(4),
                             child: Image.network(
-                              item["img"],
+                              item["img"]!,
                               fit: BoxFit.cover,
                               alignment: Alignment.topCenter,
                             ),
@@ -54,17 +54,17 @@ class _ProductCarouselState extends State<ProductCarousel> {
                       padding: const EdgeInsets.only(top: 16.0),
                       child: Column(
                         children: [
-                          Text(item["price"],
+                          Text(item["price"]!,
                               style: TextStyle(
                                   fontSize: 16, color: SunRunColors.text)),
-                          Text(item["title"],
+                          Text(item["title"]!,
                               style: TextStyle(
                                   fontSize: 32, color: SunRunColors.text)),
                           Padding(
                             padding: const EdgeInsets.only(
                                 left: 16.0, right: 16.0, top: 8),
                             child: Text(
-                              item["description"],
+                              item["description"]!,
                               style: TextStyle(
                                   fontSize: 16, color: SunRunColors.muted),
                               textAlign: TextAlign.center,
