@@ -62,8 +62,8 @@ class Home extends StatelessWidget {
                   padding: const EdgeInsets.only(top: 16.0),
                   child: CardHorizontal(
                       cta: "View article",
-                      title: homeCards["Ice Cream"]['title'],
-                      img: homeCards["Ice Cream"]['image'],
+                      title: homeCards["Ice Cream"]!['title'],
+                      img: homeCards["Ice Cream"]!['image'],
                       tap: () {
                         Navigator.pushNamed(context, '/pro');
                       }),
@@ -74,13 +74,13 @@ class Home extends StatelessWidget {
                   children: [
                     CardSmall(
                         cta: "View article",
-                        title: homeCards["Makeup"]['title'],
-                        img: homeCards["Makeup"]['image'],
+                        title: homeCards["Makeup"]!['title'],
+                        img: homeCards["Makeup"]!['image'],
                         tap: () {}),
                     CardSmall(
                         cta: "View article",
-                        title: homeCards["Coffee"]['title'],
-                        img: homeCards["Coffee"]['image'],
+                        title: homeCards["Coffee"]!['title'],
+                        img: homeCards["Coffee"]!['image'],
                         tap: () {
                           Navigator.pushNamed(context, '/pro');
                         })
@@ -89,8 +89,8 @@ class Home extends StatelessWidget {
                 SizedBox(height: 8.0),
                 CardHorizontal(
                     cta: "View article",
-                    title: homeCards["Fashion"]['title'],
-                    img: homeCards["Fashion"]['image'],
+                    title: homeCards["Fashion"]!['title'],
+                    img: homeCards["Fashion"]!['image'],
                     tap: () {
                       Navigator.pushNamed(context, '/pro');
                     }),
@@ -99,8 +99,8 @@ class Home extends StatelessWidget {
                   padding: const EdgeInsets.only(bottom: 32.0),
                   child: CardSquare(
                       cta: "View article",
-                      title: homeCards["Argon"]['title'],
-                      img: homeCards["Argon"]['image'],
+                      title: homeCards["Argon"]!['title'],
+                      img: homeCards["Argon"]!['image'],
                       tap: () {
                         Navigator.pushNamed(context, '/pro');
                       }),
